Update GenerateWorlds.bat
Fixed a bug: Sometimes world-names used a special character, which while fine for a windows-folder to use, could cause an error when loading that save in the game. So this update removes those characters before renaming the folder to that. Results are ok for now, and the gen param file still uses the full name, since it's intended for manual use anyway.