Update GenerateWorlds.bat
Big new feature! So I guess this will be the v0.4 update.
Now, as the script generates worlds, it will also output a summary of each world's populations and tower-count to a timestamped GenLog file, in the df directory. The main reason for this is to make it easier to check these numbers, all in one place. I tried to format it to make it easier to read, since after generating maps, we still have to create the legends files manually if we want more info, and these numbers sometimes help me narrow down which maps to check further.