Update GenerateWorlds.bat
Not really a bugfix or feature.. Just trying an alternative method for getting the sites and pop filename. Before, it was using the complete path for sites and pop file, which included directors above DF, so I had it use the shortened name format. Shouldn't be an issue, but I want to eliminate the possibility. So I've changed it to a more direct path/filename pointing only within the df folders.
Next version might just include some preemptive error-checking.